feat: gutter styling moved to styles
