+++ packages/edition-slices/src/slices/shared/styles.js
@@ -1,11 +1,29 @@
-import { sliceContentMaxWidth } from "@times-components/styleguide";
+import {
+  editionMaxWidth,
+  sliceContentMaxWidth,
+  spacing,
+  editionBreakpoints
+} from "@times-components/styleguide";
 
-const styles = {
+const tabletPaddingHorizontalResolver = {
+  [editionBreakpoints.medium]: spacing(2),
+  [editionBreakpoints.wide]: spacing(6),
+  [editionBreakpoints.huge]: spacing(2)
+};
+
+export default (isTablet, breakpoint) => ({
   contentWrapperStyles: {
     flex: 1,
     alignSelf: "center",
     width: sliceContentMaxWidth
-  }
-};
+  },
 
-export default styles;
+  gutterStyles: {
+    alignSelf: "center",
+    maxWidth: "100%",
+    paddingHorizontal: isTablet
+      ? tabletPaddingHorizontalResolver[breakpoint]
+      : 0,
+    width: editionMaxWidth
+  }
+});
